+++ gtars/src/tokenizers/tokens/mod.rs
@@ -38,7 +38,6 @@ impl Display for TokenizedRegion<'_> {
 /// to the regions in that universe this [TokenizedRegionSet] represents.
 pub struct TokenizedRegionSet<'a> {
     pub ids: Vec<u32>,
-    pub attention_mask: Vec<u8>,
     pub universe: &'a Universe,
 }
