+++ gtars/src/tokenizers/utils/padding.rs
@@ -1,0 +1,4 @@
+pub struct PaddingParams {
+    pub pad_id: u32,
+    pub max_length: Option<usize>,
+}
