+++ gtars/src/tokenizers/utils/truncation.rs
@@ -1,0 +1,3 @@
+pub struct TruncationParams {
+    pub max_length: usize,
+}
